add table status color
