+++ components/HackneyDS/Tag/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function Tag({ children, color, outline, noBackground, className = '' }) {
+export default function Tag({ children, color, border, outline, noBackground, className = '' }) {
   // available colors = green, gray, yellow, red, blue
   const outerClass = className ? ` ${className}` : '';
   const outlineClass = outline ? ' outline' : '';
@@ -8,6 +8,8 @@ export default function Tag({ children, color, outline, noBackground, className 
   const noBackgroundClass = noBackground ? ' no-background' : '';
 
   return (
-    <span className={`lbh-tag govuk-tag${outlineClass}${tagColorClass}${outerClass}${noBackgroundClass}`}>{children}</span>
+    <span style={{ border }} className={`lbh-tag govuk-tag${outlineClass}${tagColorClass}${outerClass}${noBackgroundClass}`}>
+      {children}
+    </span>
   );
 }
