+++ styles/globals.scss
@@ -63,6 +63,7 @@
 @import '../components/Pages/ServiceUser/Packages/style.scss';
 @import "css/container";
 @import "css/ServiceUserSearch";
+@import "css/colors";
 
 * {
   font-family: 'OpenSans', sans-serif !important;
